# go/go-gin-server/go/api_pet.go
/*
 * Swagger Petstore - OpenAPI 3.1
 *
 * This is a sample Pet Store Server based on the OpenAPI 3.1 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.  Some useful links: - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore) - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
 *
 * API version: 1.0.12
 * Contact: apiteam@swagger.io
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package petstore

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
)

type PetAPI struct {
	Store Store
}

// Post /api/v3/pet
// Add a new pet to the store.
func (api *PetAPI) AddPet(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	var pet Pet
	if !bindJSON(c, &pet) {
		return
	}
	created, err := api.Store.CreatePet(c.Request.Context(), pet)
	if err != nil {
		handleStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, created)
}

// Delete /api/v3/pet/:petId
// Deletes a pet.
func (api *PetAPI) DeletePet(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	id, ok := parseInt64Param(c, "petId")
	if !ok {
		return
	}
	if _, err := api.Store.DeletePet(c.Request.Context(), id); err != nil {
		handleStoreError(c, err)
		return
	}
	c.Status(http.StatusOK)
}

// Get /api/v3/pet/findByStatus
// Finds Pets by status.
func (api *PetAPI) FindPetsByStatus(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	statuses := queryList(c, "status")
	if len(statuses) == 0 {
		statuses = []string{"available"}
	}
	if !validateStatuses(statuses) {
		writeError(c, http.StatusBadRequest, "invalid_status", "status must be available, pending, or sold")
		return
	}
	pets, err := api.Store.FindPetsByStatus(c.Request.Context(), statuses)
	if err != nil {
		handleStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, pets)
}

// Get /api/v3/pet/findByTags
// Finds Pets by tags.
func (api *PetAPI) FindPetsByTags(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	pets, err := api.Store.FindPetsByTags(c.Request.Context(), queryList(c, "tags"))
	if err != nil {
		handleStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, pets)
}

// Get /api/v3/pet/:petId
// Find pet by identifier.
func (api *PetAPI) GetPetById(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	id, ok := parseInt64Param(c, "petId")
	if !ok {
		return
	}
	pet, err := api.Store.GetPetByID(c.Request.Context(), id)
	if err != nil {
		handleStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, pet)
}

// Put /api/v3/pet
// Update an existing pet.
func (api *PetAPI) UpdatePet(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	var pet Pet
	if !bindJSON(c, &pet) {
		return
	}
	updated, err := api.Store.UpdatePet(c.Request.Context(), pet)
	if err != nil {
		handleStoreError(c, err)
		return
	}
	c.JSON(http.StatusOK, updated)
}

// Post /api/v3/pet/:petId
// Updates a pet in the store with form data.
func (api *PetAPI) UpdatePetWithForm(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	id, ok := parseInt64Param(c, "petId")
	if !ok {
		return
	}
	name := optionalQuery(c, "name")
	status := optionalQuery(c, "status")
	updated, err := api.Store.UpdatePetFields(c.Request.Context(), id, name, status)
	if err != nil {
		handleStoreError(c, err)
		return
	}
	if !updated {
		writeNotFound(c, fmt.Sprintf("pet %d not found", id))
		return
	}
	c.Status(http.StatusOK)
}

// Post /api/v3/pet/:petId/uploadImage
// Uploads an image.
func (api *PetAPI) UploadFile(c *gin.Context) {
	if !requireStore(c, api.Store) {
		return
	}
	id, ok := parseInt64Param(c, "petId")
	if !ok {
		return
	}
	if _, err := api.Store.GetPetByID(c.Request.Context(), id); err != nil {
		handleStoreError(c, err)
		return
	}
	body, ok := readBody(c)
	if !ok {
		return
	}
	c.JSON(http.StatusOK, ApiResponse{
		Code:    http.StatusOK,
		Type:    "upload",
		Message: uploadMessage(id, c.Query("additionalMetadata"), len(body)),
	})
}