# java/helidon/src/main/java/org/openapitools/server/api/PetService.java
/*
 * Swagger Petstore - OpenAPI 3.1
 * This is a sample Pet Store Server based on the OpenAPI 3.1 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.  Some useful links: - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore) - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
 *
 * The version of the OpenAPI document: 1.0.12
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

package org.openapitools.server.api;

import org.openapitools.server.model.Error;
import java.io.File;
import java.util.List;
import org.openapitools.server.model.ModelApiResponse;
import java.util.Optional;
import org.openapitools.server.model.Pet;

import jakarta.ws.rs.*;

import java.io.InputStream;
import java.util.Map;
import java.util.List;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;

@Path("/pet")
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaHelidonServerCodegen", date = "2026-06-09T23:01:38.088546400-04:00[America/Toronto]", comments = "Generator version: 7.23.0")
public interface PetService {

    @POST
    @Consumes({ "application/json", "application/xml", "application/x-www-form-urlencoded" })
    @Produces({ "application/json", "application/xml" })
    Pet addPet(@Valid @NotNull Pet pet);

    @DELETE
    @Path("/{petId}")
    @Produces({ "application/json" })
    void deletePet(@PathParam("petId") Long petId, @HeaderParam("api_key")  String apiKey);

    @GET
    @Path("/findByStatus")
    @Produces({ "application/json", "application/xml" })
    List<Pet> findPetsByStatus(@QueryParam("status") @DefaultValue("available") String status);

    @GET
    @Path("/findByTags")
    @Produces({ "application/json", "application/xml" })
    List<Pet> findPetsByTags(@QueryParam("tags") List<String> tags);

    @GET
    @Path("/{petId}")
    @Produces({ "application/json", "application/xml" })
    Pet getPetById(@PathParam("petId") Long petId);

    @PUT
    @Consumes({ "application/json", "application/xml", "application/x-www-form-urlencoded" })
    @Produces({ "application/json", "application/xml" })
    Pet updatePet(@Valid @NotNull Pet pet);

    @POST
    @Path("/{petId}")
    @Produces({ "application/json" })
    void updatePetWithForm(@PathParam("petId") Long petId, @QueryParam("name") String name, @QueryParam("status") String status);

    @POST
    @Path("/{petId}/uploadImage")
    @Consumes({ "application/octet-stream" })
    @Produces({ "application/json" })
    ModelApiResponse uploadFile(@PathParam("petId") Long petId, @QueryParam("additionalMetadata") String additionalMetadata, @Valid File body);
}