# java/helidon/src/main/java/org/openapitools/server/api/StoreService.java
/*
 * Swagger Petstore - OpenAPI 3.1
 * This is a sample Pet Store Server based on the OpenAPI 3.1 specification.  You can find out more about Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach! You can now help us improve the API whether it's by making changes to the definition itself or to the code. That way, with time, we can improve the API in general, and expose some of the new features in OAS3.  Some useful links: - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore) - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
 *
 * The version of the OpenAPI document: 1.0.12
 * Contact: apiteam@swagger.io
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

package org.openapitools.server.api;

import org.openapitools.server.model.Error;
import java.util.Map;
import org.openapitools.server.model.Order;

import jakarta.ws.rs.*;

import java.io.InputStream;
import java.util.Map;
import java.util.List;
import jakarta.validation.constraints.*;
import jakarta.validation.Valid;

@Path("/store")
@jakarta.annotation.Generated(value = "org.openapitools.codegen.languages.JavaHelidonServerCodegen", date = "2026-06-09T23:01:38.088546400-04:00[America/Toronto]", comments = "Generator version: 7.23.0")
public interface StoreService {

    @DELETE
    @Path("/order/{orderId}")
    @Produces({ "application/json" })
    void deleteOrder(@PathParam("orderId") Long orderId);

    @GET
    @Path("/inventory")
    @Produces({ "application/json" })
    Map<String, Integer> getInventory();

    @GET
    @Path("/order/{orderId}")
    @Produces({ "application/json", "application/xml" })
    Order getOrderById(@PathParam("orderId") Long orderId);

    @POST
    @Path("/order")
    @Consumes({ "application/json", "application/xml", "application/x-www-form-urlencoded" })
    @Produces({ "application/json" })
    Order placeOrder(@Valid Order order);
}