# python/openapi.yaml
openapi: 3.1.0
info:
  contact:
    email: apiteam@swagger.io
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.1 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Swagger Petstore - OpenAPI 3.1
  version: 1.0.12
...
servers:
- url: https://petstore31.swagger.io/api/v3
...
paths:
  /pet:
    post:
      description: Add a new pet to the store.
      operationId: addPet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pet"
          application/xml:
            schema:
              $ref: "#/components/schemas/Pet"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Pet"
        description: Create a new pet in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
            application/xml:
              schema:
                $ref: "#/components/schemas/Pet"
          description: Successful operation
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Add a new pet to the store.
      tags:
      - pet
    put:
      description: Update an existing pet by Id.
      operationId: updatePet
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Pet"
          application/xml:
            schema:
              $ref: "#/components/schemas/Pet"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Pet"
        description: Update an existent pet in the store
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
            application/xml:
              schema:
                $ref: "#/components/schemas/Pet"
          description: Successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
        "422":
          description: Validation exception
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Update an existing pet.
      tags:
      - pet
  /pet/findByStatus:
    get:
      description: Multiple status values can be provided with comma separated strings.
      operationId: findPetsByStatus
      parameters:
      - description: Status values that need to be considered for filter
        explode: true
        in: query
        name: status
        required: false
        schema:
          default: available
          enum:
          - available
          - pending
          - sold
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Pet"
                type: array
                default: null
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Pet"
                type: array
                default: null
          description: successful operation
        "400":
          description: Invalid status value
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Finds Pets by status.
      tags:
      - pet
  /pet/findByTags:
    get:
      description: "Multiple tags can be provided with comma separated strings. Use\
        \ tag1, tag2, tag3 for testing."
      operationId: findPetsByTags
      parameters:
      - description: Tags to filter by
        explode: true
        in: query
        name: tags
        required: false
        schema:
          items:
            type: string
          type: array
          default: null
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/Pet"
                type: array
                default: null
            application/xml:
              schema:
                items:
                  $ref: "#/components/schemas/Pet"
                type: array
                default: null
          description: successful operation
        "400":
          description: Invalid tag value
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Finds Pets by tags.
      tags:
      - pet
  /pet/{petId}:
    delete:
      description: delete a pet.
      operationId: deletePet
      parameters:
      - description: ""
        explode: false
        in: header
        name: api_key
        required: false
        schema:
          type: string
        style: simple
      - description: Pet id to delete
        explode: false
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid pet value
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Deletes a pet.
      tags:
      - pet
    get:
      description: Returns a single pet.
      operationId: getPetById
      parameters:
      - description: ID of pet to return
        explode: false
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pet"
            application/xml:
              schema:
                $ref: "#/components/schemas/Pet"
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Pet not found
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - api_key: []
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Find pet by identifier.
      tags:
      - pet
    post:
      description: update a pet via the form data.
      operationId: updatePetWithForm
      parameters:
      - description: ID of pet that needs to be updated
        explode: false
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: Name of pet that needs to be updated
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      - description: Status of pet that needs to be updated
        explode: true
        in: query
        name: status
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          description: successfully updated
        "400":
          description: Invalid input
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Updates a pet in the store with form data.
      tags:
      - pet
  /pet/{petId}/uploadImage:
    post:
      description: Upload an image of pet.
      operationId: uploadFile
      parameters:
      - description: ID of pet to update
        explode: false
        in: path
        name: petId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: Additional Metadata
        explode: true
        in: query
        name: additionalMetadata
        required: false
        schema:
          type: string
        style: form
      requestBody:
        content:
          application/octet-stream:
            schema:
              format: binary
              type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
          description: successful operation
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - petstore_auth:
        - write:pets
        - read:pets
      summary: Uploads an image.
      tags:
      - pet
  /store/inventory:
    get:
      description: Returns a map of status codes to quantities.
      operationId: getInventory
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  format: int32
                  type: integer
          description: successful operation
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      security:
      - api_key: []
      summary: Returns pet inventories by status.
      tags:
      - store
  /store/order:
    post:
      description: Place a new order in the store.
      operationId: placeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Order"
          application/xml:
            schema:
              $ref: "#/components/schemas/Order"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Order"
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
          description: successful operation
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      summary: Place an order for a pet.
      tags:
      - store
  /store/order/{orderId}:
    delete:
      description: For valid response try integer IDs with value < 1000. Anything
        above 1000 or non-integers will generate API errors.
      operationId: deleteOrder
      parameters:
      - description: ID of the order that needs to be deleted
        explode: false
        in: path
        name: orderId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      summary: Delete purchase order by identifier.
      tags:
      - store
    get:
      description: For valid response try integer IDs with value <= 5 or > 10. Other
        values will generate exceptions.
      operationId: getOrderById
      parameters:
      - description: ID of order that needs to be fetched
        explode: false
        in: path
        name: orderId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
            application/xml:
              schema:
                $ref: "#/components/schemas/Order"
          description: successful operation
        "400":
          description: Invalid ID supplied
        "404":
          description: Order not found
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unexpected error
      summary: Find purchase order by identifier.
      tags:
      - store
...
components:
...
  schemas:
    Order:
      example:
        id: 0
        petId: 6
        quantity: 1
        shipDate: 2000-01-23T04:56:07.000+00:00
        status: placed
        complete: true
      properties:
        id:
          format: int64
          title: id
          type: integer
        petId:
          format: int64
          title: petId
          type: integer
        quantity:
          format: int32
          title: quantity
          type: integer
        shipDate:
          format: date-time
          title: shipDate
          type: string
        status:
          description: Order Status
          enum:
          - placed
          - approved
          - delivered
          title: status
          type: string
        complete:
          title: complete
          type: boolean
      title: Order
      xml:
        name: order
...
    Category:
      example:
        id: 6
        name: name
      properties:
        id:
          format: int64
          title: id
          type: integer
        name:
          title: name
          type: string
      title: Category
      xml:
        name: category
...
    Tag:
      example:
        id: 1
        name: name
      properties:
        id:
          format: int64
          title: id
          type: integer
        name:
          title: name
          type: string
      title: Tag
      xml:
        name: tag
    Pet:
      example:
        id: 0
        name: name
        category:
          id: 6
          name: name
        photoUrls:
        - photoUrls
        - photoUrls
        tags:
        - id: 1
          name: name
        - id: 1
          name: name
        status: available
      properties:
        id:
          format: int64
          title: id
          type: integer
        name:
          title: name
          type: string
        category:
          $ref: "#/components/schemas/Category"
        photoUrls:
          items:
            type: string
            xml:
              name: photoUrl
          title: photoUrls
          type: array
          xml:
            wrapped: true
          default: null
        tags:
          items:
            $ref: "#/components/schemas/Tag"
          title: tags
          type: array
          xml:
            wrapped: true
          default: null
        status:
          description: pet status in the store
          enum:
          - available
          - pending
          - sold
          title: status
          type: string
      required:
      - name
      - photoUrls
      title: Pet
      xml:
        name: pet
    ApiResponse:
      example:
        code: 0
        type: type
        message: message
      properties:
        code:
          format: int32
          title: code
          type: integer
        type:
          title: type
          type: string
        message:
          title: message
          type: string
      title: ApiResponse
      xml:
        name: "##default"
    Error:
      example:
        code: code
        message: message
      properties:
        code:
          title: code
          type: string
        message:
          title: message
          type: string
      required:
      - code
      - message
      title: Error
  securitySchemes:
    petstore_auth:
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
      type: oauth2
    api_key:
      in: header
      name: api_key
      type: apiKey